+++ src/Sulu/Bundle/ContactBundle/Entity/Contact.php
@@ -15,6 +15,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use JMS\Serializer\Annotation\Accessor;
 use JMS\Serializer\Annotation\Exclude;
+use JMS\Serializer\Annotation\Expose;
 use JMS\Serializer\Annotation\Groups;
 use JMS\Serializer\Annotation\SerializedName;
 use JMS\Serializer\Annotation\VirtualProperty;
@@ -33,6 +34,8 @@ class Contact extends ApiEntity implements ContactInterface, AuditableInterface
 {
     /**
      * @var int
+     * @Expose
+     * @Groups({"frontend", "partialContact", "fullContact"})
      */
     protected $id;