+++ src/Sulu/Bundle/ContentBundle/Command/PHPCRInitCommand.php
@@ -47,6 +47,7 @@ class PHPCRInitCommand extends ContainerAwareCommand
         // init node namespace and types
         $output->writeln('Register namespace');
         $workspace->getNamespaceRegistry()->registerNamespace('sulu', 'http://sulu.io/phpcr');
+        $workspace->getNamespaceRegistry()->registerNamespace('sec', 'http://sulu.io/phpcr/sec');
         $workspace->getNamespaceRegistry()->registerNamespace(
             $this->getContainer()->getParameter('sulu.content.language.namespace'),
             'http://sulu.io/phpcr/locale'